% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_in_xl.R
\name{view_in_xl}
\alias{view_in_xl}
\title{Open a data.frame in 'Excel'}
\usage{
view_in_xl(df = NULL)
}
\arguments{
\item{df}{the name of \code{data.frame}, if not provided, an addin is launched.}
}
\description{
Open a data.frame in 'Excel'
}
\examples{
\dontrun{

# Launch addin
view_in_xl()

# or pass the name of a data.frame
view_in_xl("iris")

}
}
